refactor: update entities with eslint rules
