+++ src/entities/category/api/categoriesApi.ts
@@ -1,6 +1,6 @@
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
-import { CategoriesApiResponse } from "..";
 import { NEWS_API_KEY, NEWS_BASE_API_URL } from "@/shared/config";
+import { CategoriesApiResponse } from "../model";
 
 export const categoriesApi = createApi({
   reducerPath: "categoriesApi",
