+++ src/entities/category/api/index.ts
@@ -1,0 +1,1 @@
+export * from "./categoriesApi";
