+++ src/entities/news/model/index.ts
@@ -1,0 +1,2 @@
+export * from "./newsSlice";
+export * from "./types";
