+++ src/entities/news/ui/NewsCard/NewsCard.tsx
@@ -1,16 +1,16 @@
-import { formatTimeAgo } from "@/shared/helpers/formatTimeAgo";
-import { INews } from "../..";
-import styles from "./styles.module.css";
-import Image from "@/shared/ui/Image/Image";
 import { ReactNode } from "react";
+import { formatTimeAgo } from "@/shared/lib/date";
+import { Image } from "@/shared/ui/Image";
+import { INews } from "../../model";
+import styles from "./styles.module.css";
 
 interface Props {
   item: INews;
   type: "banner" | "item";
-  viewNewsSlot?: (news: INews) => ReactNode;
+  viewNewsSlot?: (_news: INews) => ReactNode;
 }
 
-const NewsCard = ({ item, type = "item", viewNewsSlot }: Props) => {
+export const NewsCard = ({ item, type = "item", viewNewsSlot }: Props) => {
   return (
     <li className={`${styles.card} ${type === "banner" && styles.banner}`}>
       {type === "banner" ? (
@@ -33,5 +33,3 @@ const NewsCard = ({ item, type = "item", viewNewsSlot }: Props) => {
     </li>
   );
 };
-
-export default NewsCard;
