+++ src/entities/news/ui/NewsCard/index.ts
@@ -1,0 +1,1 @@
+export { NewsCard } from "./NewsCard";
