+++ src/entities/news/ui/NewsDetails/NewsDetails.tsx
@@ -1,13 +1,13 @@
-import { formatTimeAgo } from "@/shared/helpers/formatTimeAgo";
-import { INews } from "../..";
+import { formatTimeAgo } from "@/shared/lib/date";
+import { Image } from "@/shared/ui/Image";
+import { INews } from "../../model";
 import styles from "./styles.module.css";
-import Image from "@/shared/ui/Image/Image";
 
 interface Props {
   item: INews;
 }
 
-const NewsDetails = ({ item }: Props) => {
+export const NewsDetails = ({ item }: Props) => {
   return (
     <div className={styles.details}>
       <Image image={item.image} />
@@ -15,7 +15,7 @@ const NewsDetails = ({ item }: Props) => {
       <div className={styles.description}>
         <p>
           {item.description} ({item.language}){" "}
-          <a target="_blank" href={item.url}>
+          <a target="_blank" rel="noreferrer" href={item.url}>
             Read more...
           </a>
         </p>
@@ -25,12 +25,14 @@ const NewsDetails = ({ item }: Props) => {
 
         <ul>
           {item.category.map((category) => {
-            return <button className={styles.active}>{category}</button>;
+            return (
+              <button key={category} className={styles.active}>
+                {category}
+              </button>
+            );
           })}
         </ul>
       </div>
     </div>
   );
 };
-
-export default NewsDetails;
