+++ src/entities/news/ui/NewsDetails/index.ts
@@ -1,0 +1,1 @@
+export { NewsDetails } from "./NewsDetails";
