+++ src/entities/news/ui/index.ts
@@ -1,0 +1,2 @@
+export { NewsCard } from "./NewsCard";
+export { NewsDetails } from "./NewsDetails";
